pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "ptool"
dependencies = [
    "pillow"
]
version = "2025.10"

[project.optional-dependencies]
heif = [
    "pillow-heif"
]
dev = [
    "ruff",
    "mypy"
]

[project.scripts]
ptool = "ptool.__main__:entry"

[tool.basedpyright]
reportUnusedCallResult = false
reportAny = false

[tool.mypy]
disable_error_code = ["import-untyped"]

[tool.ruff]
line-length = 100


[tool.ruff.lint]
dummy-variable-rgx = "^_$"
select = [
    "E",     # pycodestyle ┐
    "F",     # pyflakes    ├ ~= flake8
    "W",     # warnings    ┘
    "I",     # isort
    "UP",    # pyupgrade
    "A",     # flake8-builtins
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "LOG",   # flake8-logging
    "G",     # flake8-logging-format
    "YTT",   # flake8-2020
    "FBT",   # flake8-boolean-trap
    "C4",    # flake8-comprehensions
]
